Go back to most of the methods being in the translatable module
